moved loadFriends() to Friends; hid the contact search due to bug
